+++ lib/app_date_picker.dart
@@ -1,5 +1,1 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'date_time_picker.dart';